requirements.txt
# Django и основные зависимости
Django==4.2.16
django-environ==0.11.2

# База данных
psycopg2-binary==2.9.9

# Кэширование и очереди
redis==5.0.1
celery==5.3.4
django-redis==5.4.0

# Статические файлы и медиа
Pillow==10.1.0
django-cleanup==8.0.0

# Формы и админка
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-admin-interface==0.28.6

# API и HTMX
django-cors-headers==4.3.1
django-htmx==1.17.0

# Аналитика
django-analytical==3.1.0

# Утилиты
python-slugify==8.0.1
django-mptt==0.15.0
django-extensions==3.2.3
python-decouple==3.8

# Локализация
django-modeltranslation==0.18.11

# Безопасность
django-ratelimit==4.1.0

# Разработка
django-debug-toolbar==4.2.0
factory-boy==3.3.0

# Продакшен
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0

# Тестирование
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0